backend/requirements.txt
fastapi
uvicorn[standard]
stripe
python-dotenv
supabase
